+++ projects/demo/src/app/app.config.ts
@@ -1,10 +1,10 @@
 import { ApplicationConfig, provideZoneChangeDetection } from '@angular/core';
-import { provideNgxLoadingOverlay } from '@shaman-apprentice/ngx-loading-overlay';
-import { LoadingOverlayComponent } from './components/loadingOverlay.component';
+import { provideNgxLoadingIndicator } from '@shaman-apprentice/ngx-loading-overlay';
+import { LoadingIndicatorComponent } from './components/loadingIndicator.component';
 
 export const appConfig: ApplicationConfig = {
   providers: [
     provideZoneChangeDetection({ eventCoalescing: true }),
-    provideNgxLoadingOverlay(LoadingOverlayComponent),
+    provideNgxLoadingIndicator(LoadingIndicatorComponent),
   ],
 };