+++ projects/ngx-loading-overlay/src/public-api.ts
@@ -1,3 +1,3 @@
 export { IsLoadingDirective } from "./lib/isLoading.directive";
-export { provideNgxLoadingOverlay } from "./lib/loadingOverlay.token";
+export { provideNgxLoadingIndicator } from "./lib/loadingIndicator.token";